Remove email capture feature
